feat: increase the value handled by Number Base Converter to 128 bits

+++ Sources/Pages/Converters/NumberBaseConverterViewState.swift
@@ -6,7 +6,7 @@ final class NumberBaseConverterViewState {
         didSet { self.formatAllText() }
     }
     var inputType = NumberType.decimal
-    private(set) var inputValue: Int? {
+    private(set) var inputValue: Int128? {
         didSet { self.updateOutputValues() }
     }
     var hexadecimal = "" {
